backend/requirements.txt
flask
python-dotenv
pinecone
google-generativeai
cohere
groq
langchain-text-splitters  # Just for a robust text splitter
flask-cors # To allow our frontend to call the backend
node
PyMuPDF